refactor+fix: Extract <.alert_group /> to its own component and fix alert placement
